+++ History.md
@@ -7,6 +7,7 @@
 * Moved all runner code into the mirror
 * Added experimental parallel testing mode
 * Includes a DDP connection to the mirror by default
+* Sample tests are much simpler now with a fixture, ddp and webdriver example
 
 #0.5.5